Move field array file to step directory
