+++ packages/js/src/workouts/tailwind-components/steps/social-profiles/social-input-section.js
@@ -2,7 +2,7 @@ import { __ } from "@wordpress/i18n";
 import { useCallback } from "@wordpress/element";
 import PropTypes from "prop-types";
 
-import FieldArray from "../../base/field-array";
+import FieldArray from "./field-array";
 import SocialInput from "./social-input";
 
 /* eslint-disable complexity */
